feat: add scanning button in units and pages fragment, remove scanner fragment
