+++ app/src/main/java/com/example/smartnote/fragments/PagesFragment.kt
@@ -1,11 +1,21 @@
 package com.example.smartnote.fragments
 
+import android.Manifest
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.graphics.Bitmap
+import android.graphics.ImageDecoder
+import android.net.Uri
+import android.os.Build
 import android.os.Bundle
+import android.provider.MediaStore
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.navArgs
@@ -14,10 +24,14 @@ import com.example.smartnote.MainActivity
 import com.example.smartnote.adapters.PagesAdapter
 import com.example.smartnote.databinding.FragmentPagesBinding
 import com.example.smartnote.db.Pdf
+import com.example.smartnote.helpers.Constants
 import com.example.smartnote.helpers.viewLifecycle
 import com.example.smartnote.viewmodels.BookViewModel
 import com.example.smartnote.viewmodels.FileViewModel
+import com.scanlibrary.ScanActivity
+import com.scanlibrary.ScanConstants
 import dagger.hilt.android.AndroidEntryPoint
+import java.io.IOException
 import java.net.URLConnection
 import java.util.*
 
@@ -31,6 +45,7 @@ class PagesFragment : Fragment() {
   private val bookViewModel: BookViewModel by lazy {
     ViewModelProvider(requireActivity()).get(BookViewModel::class.java)
   }
+  private val args: PagesFragmentArgs by navArgs()
 
   override fun onCreateView(
     inflater: LayoutInflater,
@@ -44,7 +59,9 @@ class PagesFragment : Fragment() {
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
-    val args: PagesFragmentArgs by navArgs()
+    binding.buttonAddPhoto.setOnClickListener {
+      checkPermissionAndScan()
+    }
     (activity as MainActivity).supportActionBar?.title = "${args.subjectName} - Unit ${args.unitNo}"
     val fileStrings = mutableListOf<String>()
     val list = context?.let { viewModel.getFiles(args.unitFolderPath, it) }
@@ -90,6 +107,59 @@ class PagesFragment : Fragment() {
         bookViewModel.deletePdfByName(args.unitFolderPath.split('/').toString())
         bookViewModel.insertPdf(pdf)
         Toast.makeText(activity, "Pdf created successfully", Toast.LENGTH_SHORT).show()
+      }
+    }
+  }
+
+  private fun checkPermissionAndScan() {
+    if (ActivityCompat.checkSelfPermission(
+        requireContext(),
+        Manifest.permission.CAMERA
+      ) != PackageManager.PERMISSION_GRANTED ||
+      ActivityCompat.checkSelfPermission(
+        requireContext(),
+        Manifest.permission.WRITE_EXTERNAL_STORAGE
+      ) != PackageManager.PERMISSION_GRANTED
+    ) {
+      requestPermissions(
+        arrayOf(
+          Manifest.permission.CAMERA,
+          Manifest.permission.WRITE_EXTERNAL_STORAGE
+        ),
+        Constants.PERMISSIONS_REQUEST_CODE
+      )
+    } else {
+      Log.i("SubjectsFragment", "Permissions Granted")
+      val preference = ScanConstants.OPEN_CAMERA
+      val intent = Intent(requireContext(), ScanActivity::class.java)
+      intent.putExtra(ScanConstants.OPEN_INTENT_PREFERENCE, preference)
+      startActivityForResult(intent, Constants.SCANNER_REQUEST_CODE)
+    }
+  }
+
+  override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+    super.onActivityResult(requestCode, resultCode, data)
+    if (requestCode == Constants.SCANNER_REQUEST_CODE && resultCode == AppCompatActivity.RESULT_OK) {
+      val uri: Uri = data!!.extras!!.getParcelable(ScanConstants.SCANNED_RESULT)!!
+      var bitmap: Bitmap? = null
+      val contentResolver = requireActivity().contentResolver
+      try {
+        bitmap = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+          val source = ImageDecoder.createSource(contentResolver, uri)
+          ImageDecoder.decodeBitmap(source)
+        } else {
+          MediaStore.Images.Media.getBitmap(contentResolver, uri)
+        }
+        contentResolver.delete(uri, null, null)
+
+        val scannedSub = data.extras!!.get("scannedSub")
+        val scannedUnit = data.extras!!.get("scannedUnit")
+
+        val path = bookViewModel.getSubjectFolderPath(args.bookName, subNo = scannedSub as Int)
+
+        viewModel.storeImage(bitmap, scannedUnit.toString(), path)
+      } catch (e: IOException) {
+        e.printStackTrace()
       }
     }
   }
