+++ app/src/main/java/com/example/smartnote/fragments/SubjectsFragment.kt
@@ -1,9 +1,20 @@
 package com.example.smartnote.fragments
 
+import android.Manifest
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.graphics.Bitmap
+import android.graphics.ImageDecoder
+import android.net.Uri
+import android.os.Build
 import android.os.Bundle
+import android.provider.MediaStore
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
@@ -16,9 +27,15 @@ import com.example.smartnote.R
 import com.example.smartnote.adapters.SubjectsAdapter
 import com.example.smartnote.databinding.FragmentSubjectsBinding
 import com.example.smartnote.db.Book
+import com.example.smartnote.helpers.Constants.PERMISSIONS_REQUEST_CODE
+import com.example.smartnote.helpers.Constants.SCANNER_REQUEST_CODE
 import com.example.smartnote.helpers.viewLifecycle
 import com.example.smartnote.viewmodels.BookViewModel
+import com.example.smartnote.viewmodels.FileViewModel
+import com.scanlibrary.ScanActivity
+import com.scanlibrary.ScanConstants
 import dagger.hilt.android.AndroidEntryPoint
+import java.io.IOException
 
 @AndroidEntryPoint
 class SubjectsFragment : Fragment() {
@@ -29,6 +46,10 @@ class SubjectsFragment : Fragment() {
   private val viewModel: BookViewModel by lazy {
     ViewModelProvider(requireActivity()).get(BookViewModel::class.java)
   }
+  private val fileViewModel: FileViewModel by lazy {
+    ViewModelProvider(this).get(FileViewModel::class.java)
+  }
+
 
   // recycler_view
   private lateinit var recyclerView: RecyclerView
@@ -51,7 +72,7 @@ class SubjectsFragment : Fragment() {
       val bookName: String = args.bookName
       val bundle = bundleOf("bookName" to bookName)
       binding.buttonAddPhoto.setOnClickListener {
-        it.findNavController().navigate(R.id.action_subjectsFragment_to_scannerFragment, bundle)
+        checkPermissionAndScan()
       }
       adapter = SubjectsAdapter(book)
       viewModel.getBookById(bookId).observe(
@@ -66,6 +87,59 @@ class SubjectsFragment : Fragment() {
       recyclerView = binding.recyclerView
       recyclerView.layoutManager = LinearLayoutManager(activity)
       recyclerView.adapter = adapter
+    }
+  }
+
+  private fun checkPermissionAndScan() {
+    if (ActivityCompat.checkSelfPermission(
+        requireContext(),
+        Manifest.permission.CAMERA
+      ) != PackageManager.PERMISSION_GRANTED ||
+      ActivityCompat.checkSelfPermission(
+        requireContext(),
+        Manifest.permission.WRITE_EXTERNAL_STORAGE
+      ) != PackageManager.PERMISSION_GRANTED
+    ) {
+      requestPermissions(
+        arrayOf(
+          Manifest.permission.CAMERA,
+          Manifest.permission.WRITE_EXTERNAL_STORAGE
+        ),
+        PERMISSIONS_REQUEST_CODE
+      )
+    } else {
+      Log.i("SubjectsFragment", "Permissions Granted")
+      val preference = ScanConstants.OPEN_CAMERA
+      val intent = Intent(requireContext(), ScanActivity::class.java)
+      intent.putExtra(ScanConstants.OPEN_INTENT_PREFERENCE, preference)
+      startActivityForResult(intent, SCANNER_REQUEST_CODE)
+    }
+  }
+
+  override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+    super.onActivityResult(requestCode, resultCode, data)
+    if (requestCode == SCANNER_REQUEST_CODE && resultCode == AppCompatActivity.RESULT_OK) {
+      val uri: Uri = data!!.extras!!.getParcelable(ScanConstants.SCANNED_RESULT)!!
+      var bitmap: Bitmap? = null
+      val contentResolver = requireActivity().contentResolver
+      try {
+        bitmap = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+          val source = ImageDecoder.createSource(contentResolver, uri)
+          ImageDecoder.decodeBitmap(source)
+        } else {
+          MediaStore.Images.Media.getBitmap(contentResolver, uri)
+        }
+        contentResolver.delete(uri, null, null)
+
+        val scannedSub = data.extras!!.get("scannedSub")
+        val scannedUnit = data.extras!!.get("scannedUnit")
+
+        val path = viewModel.getSubjectFolderPath(bookName = requireArguments().getString("bookName", ""), subNo = scannedSub as Int)
+
+        fileViewModel.storeImage(bitmap, scannedUnit.toString(), path)
+      } catch (e: IOException) {
+        e.printStackTrace()
+      }
     }
   }
 }
