+++ app/src/main/java/com/example/smartnote/helpers/Constants.kt
@@ -4,4 +4,6 @@ object Constants {
     const val TYPE_FOLDER = "application/vnd.google-apps.folder"
     const val TYPE_PDF = "application/pdf"
     const val LAST_SYNCED_TIME = "lastSyncedTime"
+    const val SCANNER_REQUEST_CODE = 99
+    const val PERMISSIONS_REQUEST_CODE = 9
 }
