+++ app/src/main/res/layout/fragment_pages.xml
@@ -10,13 +10,31 @@
         android:id="@+id/pagesRecyclerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/button_create_pdf"
+
+    <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="end|bottom"
-        android:src="@drawable/ic_baseline_picture_as_pdf_24"
-        android:layout_margin="16dp"
-        android:contentDescription="Create PDF" />
+        android:orientation="vertical"
+        android:layout_gravity="end|bottom">
+
+        <com.google.android.material.floatingactionbutton.FloatingActionButton
+            android:id="@+id/button_create_pdf"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="end|bottom"
+            android:src="@drawable/ic_baseline_picture_as_pdf_24"
+            android:layout_marginEnd="16dp"
+            android:contentDescription="Create PDF" />
+
+        <com.google.android.material.floatingactionbutton.FloatingActionButton
+            android:id="@+id/button_add_photo"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="end|bottom"
+            android:src="@drawable/ic_baseline_add_a_photo_24"
+            android:layout_margin="16dp"
+            android:contentDescription="Add Pages" />
+
+    </LinearLayout>
 
 </FrameLayout>
