+++ app/src/main/res/layout/fragment_units.xml
@@ -9,4 +9,14 @@
         android:id="@+id/unitsRecyclerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/button_add_photo"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end|bottom"
+        android:src="@drawable/ic_baseline_add_a_photo_24"
+        android:layout_margin="16dp"
+        android:contentDescription="Add Pages" />
+
 </FrameLayout>
